Added bottom animations
