+++ client/src/components/findJobs/index (DESKTOP-ANEJHLO's conflicted copy 2020-02-05).js
@@ -1,0 +1,17 @@
+import React, { Component } from 'react';
+import { Redirect } from 'react-router-dom';
+import axios from 'axios'
+import './style.css';
+
+
+//Broker and Carrier Dash
+
+
+class findJobs extends Component {
+    constructor(props) {
+        super(props);
+        this.state = {
+            
+        }
+    }
+}
