+++ client/src/components/footer/index (DESKTOP-ANEJHLO's conflicted copy 2020-02-05).js
@@ -1,0 +1,3 @@
+import React from 'react';
+import './style.css';
+
