+++ client/src/components/supplyJobs/index (DESKTOP-ANEJHLO's conflicted copy 2020-02-05).js
@@ -1,0 +1,4 @@
+import React from 'react';
+import './style.css'
+
+//Create a Job on the Supplier Tab
