+++ client/src/pages/SignUp.js
@@ -1,4 +1,102 @@
-import React, { Component } from 'react';
-import Nav from '../components/navLogin';
-import Form from '../components/signUpForm';
-import Footer from '../components/footer';
+import React, { Component } from "react";
+import { withRouter } from 'react-router'
+import axios from "axios";
+
+class Signup extends Component {
+    state = {
+      username: "",
+      password: "",
+      confirmPassword: ""
+    };
+
+  
+  handleChange = (event) => {
+    this.setState({
+      [event.target.name]: event.target.value
+    });
+  }
+
+  handleSubmit = (event) => {
+    event.preventDefault();
+
+    //request to server to add a new username/password
+    console.log('handleSubmit Object:', { username: this.state.username, password: this.state.password });
+    axios
+      .post("/api/user/", {
+        username: this.state.username,
+        password: this.state.password
+      })
+      .then(response => {
+        console.log(response);
+        if (!response.data.errmsg) {
+          console.log("successful signup");
+          //Redirect to login page
+          this.props.history.push('/login', {user: this.state.username})
+        } else {
+          console.log("username already taken");
+        }
+      })
+      .catch(error => {
+        console.log("signup error: ");
+        console.log(error);
+      });
+  }
+
+  render() {
+    console.log('doesPropsExist:', this.props);
+    return (
+      <div className="SignupForm">
+        <h4>Sign up</h4>
+        <form className="form-horizontal">
+          <div className="form-group">
+            <div className="col-1 col-ml-auto">
+              <label className="form-label" htmlFor="username">
+                Username
+              </label>
+            </div>
+            <div className="col-3 col-mr-auto">
+              <input
+                className="form-input"
+                type="text"
+                id="username"
+                name="username"
+                placeholder="Username"
+                value={this.state.username}
+                onChange={this.handleChange}
+              />
+            </div>
+          </div>
+          <div className="form-group">
+            <div className="col-1 col-ml-auto">
+              <label className="form-label" htmlFor="password">
+                Password:{" "}
+              </label>
+            </div>
+            <div className="col-3 col-mr-auto">
+              <input
+                className="form-input"
+                placeholder="password"
+                type="password"
+                name="password"
+                value={this.state.password}
+                onChange={this.handleChange}
+              />
+            </div>
+          </div>
+          <div className="form-group ">
+            <div className="col-7"></div>
+            <button
+              className="btn btn-primary col-1 col-mr-auto"
+              onClick={this.handleSubmit}
+              type="submit"
+            >
+              Sign up
+            </button>
+          </div>
+        </form>
+      </div>
+    );
+  }
+}
+
+export default withRouter(Signup);
