+++ database/index.js
@@ -4,7 +4,7 @@ mongoose.Promise = global.Promise;
 
 //your local database url
 //27017 is the default mongoDB port
-const uri = "mongodb://localhost:27017/camion";
+const uri = process.env.MONGODB_URI || "mongodb://localhost:27017/camion";
 
 mongoose.connect(uri).then(
   () => {
