+++ package.json
@@ -37,6 +37,8 @@
     "mongoose": "^5.8.9",
     "morgan": "^1.9.1",
     "passport": "^0.4.1",
-    "passport-local": "^1.0.0"
-  }
+    "passport-local": "^1.0.0",
+    "react-scripts": "^3.3.1"
+  },
+  "devDependencies": {}
 }
