+++ routes/api/index (DESKTOP-ANEJHLO's conflicted copy 2020-02-05).js
@@ -1,0 +1,7 @@
+const router = require("express").Router();
+const userRoutes = require("./user");
+
+// User routes
+router.use("/user", userRoutes);
+
+module.exports = router;
